Fix Runtime error on first launch cuz me dumb

+++ app/assets/js/scripts/landing.js
@@ -1098,7 +1098,7 @@ function dlAsync(login = true){
                                     fs.readdirSync(paths.mods).forEach((file) => {
                                         // Prevent optifine to be deleted here because of Java Path issues
                                         // Shit patch but honestly I don't care, I don't have time to implement something better
-                                        if(file !== 'OptiFine.jar' && file !== 'MixinBootstrap.jar') {
+                                        if(file !== 'OptiFine.jar' && file !== 'MixinBootstrap.jar' && file !== "nicephore") {
                                             fs.unlinkSync(path.join(paths.mods, file))
                                         }
                                     })
